# Modification of Pixel Pair Step in HI Tracking