--- a/RecoHI/HiTracking/python/hiPixelPairStep_cff.py
+++ b/RecoHI/HiTracking/python/hiPixelPairStep_cff.py
@@ -40,7 +40,11 @@ hiPixelPairSeedLayers = RecoTracker.TkSeedingLayers.PixelLayerPairs_cfi.PixelLay
         HitProducer = cms.string('siPixelRecHits'),
         skipClusters = cms.InputTag('hiPixelPairClusters')
         )
-            )
+)
+from Configuration.Eras.Modifier_trackingPhase1_cff import trackingPhase1
+trackingPhase1.toModify(hiPixelPairSeedLayers,
+	layerList = cms.vstring('BPix1+BPix4','BPix1+FPix1_pos','BPix1+FPix1_neg')  #only use first and fourth barrel layers or first barrel and first forward layer around area where BPIX2+3 are inactive
+)
 
 # SEEDS
 import RecoTracker.TkSeedGenerator.GlobalSeedsFromPairsWithVertices_cff
@@ -67,6 +71,47 @@ hiPixelPairSeeds.SeedComparitorPSet = cms.PSet(
     ClusterShapeHitFilterName = cms.string('ClusterShapeHitFilter'),
     ClusterShapeCacheSrc = cms.InputTag("siPixelClusterShapeCache")
     )
+
+#rectangular tracking region around area missing BPIX2/3 in Phase 1
+from RecoTracker.TkTrackingRegions.pointSeededTrackingRegion_cfi import pointSeededTrackingRegion as _pointSeededTrackingRegion
+hiPixelPairStepTrackingRegionPhase1 = _pointSeededTrackingRegion.clone(
+    RegionPSet = dict(
+        ptMin = 0.9,
+        originRadius = 0.005,
+        mode = "VerticesSigma",
+        nSigmaZVertex = 4.0,
+        vertexCollection = "hiSelectedPixelVertex",
+        beamSpot = "offlineBeamSpot",
+        whereToUseMeasurementTracker = "Never",
+        deltaEta = 1.8,
+        deltaPhi = 0.5,
+        points = dict(
+            eta = [0.0],
+            phi = [3.0],
+        )
+    )
+)
+
+from RecoTracker.TkHitPairs.hitPairEDProducer_cfi import hitPairEDProducer as _hitPairEDProducer
+hiPixelPairStepHitDoubletsPhase1 = _hitPairEDProducer.clone(
+    seedingLayers = "hiPixelPairSeedLayers",
+    trackingRegions = "hiPixelPairStepTrackingRegionPhase1",
+    clusterCheck = "",
+    produceSeedingHitSets = cms.bool(True), 
+)
+
+from RecoTracker.TkSeedGenerator.seedCreatorFromRegionConsecutiveHitsEDProducer_cff import seedCreatorFromRegionConsecutiveHitsEDProducer as _seedCreatorFromRegionConsecutiveHitsEDProducer
+hiPixelPairStepSeedsPhase1 = _seedCreatorFromRegionConsecutiveHitsEDProducer.clone(
+    seedingHitSets = "hiPixelPairStepHitDoubletsPhase1",
+    SeedComparitorPSet = dict(
+        ComponentName = 'PixelClusterShapeSeedComparitor',
+        FilterAtHelixStage = cms.bool(True),
+        FilterPixelHits = cms.bool(True),
+        FilterStripHits = cms.bool(False),
+        ClusterShapeHitFilterName = cms.string('ClusterShapeHitFilter'),
+        ClusterShapeCacheSrc = cms.InputTag('siPixelClusterShapeCache'),
+    )
+)
 
 # QUALITY CUTS DURING TRACK BUILDING
 import TrackingTools.TrajectoryFiltering.TrajectoryFilter_cff
@@ -105,6 +150,9 @@ hiPixelPairTrackCandidates = RecoTracker.CkfPattern.CkfTrackCandidates_cfi.ckfTr
     onlyPixelHitsForSeedCleaner = cms.bool(True),
 
     )
+trackingPhase1.toModify(hiPixelPairTrackCandidates,
+    src = cms.InputTag('hiPixelPairStepSeedsPhase1')
+)
 
 
 # TRACK FITTING
@@ -168,10 +216,12 @@ trackingPhase1.toModify(hiPixelPairStepSelector, trackSelectors= cms.VPSet(
 
 
 # Final sequence
-
 hiPixelPairStep = cms.Sequence(hiPixelPairClusters*
                                hiPixelPairSeedLayers*
                                hiPixelPairSeeds*
                                hiPixelPairTrackCandidates*
                                hiPixelPairGlobalPrimTracks*
                                hiPixelPairStepSelector)
+hiPixelPairStep_Phase1 = hiPixelPairStep.copy()
+hiPixelPairStep_Phase1.replace(hiPixelPairSeeds,hiPixelPairStepTrackingRegionPhase1*hiPixelPairStepHitDoubletsPhase1*hiPixelPairStepSeedsPhase1)
+trackingPhase1.toReplaceWith(hiPixelPairStep, hiPixelPairStep_Phase1)