After the inclusion of additional quadruplet iterations for Phase 1 tracking in PR #18646, we have decided to mostly remove the pixel pair step from the phase 1 HI tracking workflow (similar to what was done for pp tracking). Studies here show that this iteration is only recovering 1-2% efficiency, while also contributing ~1% fake rate.  A small tracking region for pixel layers 1 and 4, and forward pixel disk 1 has been kept to cover a gap in acceptance of BPIX2 and 3 (very similar to this pp PR: #19051). 

Timing studies show an overall decrease in tracking time in 0-10% events from ~109 s to ~99 s, and from 10.3 to 9.7s in inclusive events.

Testing can be done using the HI Phase 1 relval proposed in PR #19089.
@mandrenguyen