@cmsbuild 2017-06-29
The tests are being triggered in jenkins.