@cmsbuild 2017-06-29
+1
Tested at: 67388259348f93fd96ff323121cdae9af4f7a6ae
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-19486/21001/summary.html

The following merge commits were also included on top of IB + this PR after doing git cms-merge-topic: 
https://github.com/cms-sw/cmssw/commit/1974578bdc7b4ef3a501f360274c7bec46ac7066
You can see more details here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-19486/21001/git-log-recent-commits
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-19486/21001/git-merge-result