@perrotta 2017-07-04
+1
- Configuration changes for hiPixelPairStep: mostly removing pixel pair step from HI tracking in Phase 1
- As stated in the description, it leads to an overall reduction of the HI tracking time.
- Tested running with the newly defined workflows 150,, 150.1, 150.2, 150.3
- Performance of the modified HI tracking was studied by the proponents and summarized in the PR description