@perrotta 2017-07-04
I removed the "+1" from the comment above, not because I found anything wrong with this PR, but because I would like to see some of those performance plots (eff, fakerates, ...) you summarizes in the PR description.

Could you please provide a link to a presentation in which you showed those performances, or just attach some distribution here? Then "+1" will follow.