@abaty 2017-07-04
@perrotta 
A study of the efficiency and fake rate from this iteration can be found on slide 4,5 [here](https://indico.cern.ch/event/646309/contributions/2625076/attachments/1475706/2285895/PixelPairStudy.pdf).
Justification for the regional tracking around the hole in BPIX2+3, as well as timing numbers can be found [here](https://indico.cern.ch/event/649262/contributions/2639830/attachments/1483753/2302322/Baty_PixelPairUpdate_TrackingMeeting.pdf).