@perrotta 2017-07-05
Thank you @abaty 

As far as I understand, in https://indico.cern.ch/event/646309/contributions/2625076/attachments/1475706/2285895/PixelPairStudy.pdf you have the efficiency (and fake-rates) for including the pixelPairStep everywhere.

Since in this PR you only apply it in a eta/phi slice, do you have the same plots with the same config as in this PR? Page 6 in https://indico.cern.ch/event/649262/contributions/2639830/attachments/1483753/2302322/Baty_PixelPairUpdate_TrackingMeeting.pdf makes me confident that the efficiency is correctly recovered where needed, but it would be nice to have here also plots with the actual setup proposed here.