@abaty 2017-07-05
@perrotta 

You can find the efficiency and fake rates for the code in this PR (i.e. only the regional pair step) [here](https://twiki.cern.ch/twiki/pub/CMS/MITHIGAnalysis2017/Effect_Of_PixelPair_PR19486.pdf).  The extra tracks found by the pair step are shown by the dark brown contributions to the histogram.