@perrotta 2017-07-05
+1
- Configuration changes for hiPixelPairStep: it removes pixel pair step from HI tracking in Phase 1 everywhere but in a slice needed to recover efficiency for holes in  BPix2+3 layers
- As stated in the description, it leads to an overall reduction of the HI tracking time with respect to the usage of PixelPairStep everywhere; about 4 to 9% of that reduction (depending on centrality) is given back when the pixelPairStep is only used in the slice
- Tested running with the newly defined workflows 150,, 150.1, 150.2, 150.3
- Performance of the modified HI tracking was studied by the proponents and summarized in the PR description, and in https://github.com/cms-sw/cmssw/pull/19486#issuecomment-313099682